tabel matkul

+++ database/migrations/4_create_mahasiswa_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::create('mahasiswa', function (Blueprint $table) {
-            $table->integer('NIM')->primary;
+            $table->bigIncrements('NIM')->primary;
             $table->string('Nama_mahasiswa');
             $table->unsignedBigInteger('Dosen_NIP')->nullable();
             $table->foreign('Dosen_NIP')->references('NIP')->on('dosen');
